Minor changes in constants and data types

+++ include/cycloid.h
@@ -14,7 +14,7 @@ struct Cycloid
     double *radius;
 
     double *thetas;
-    double *omegas;
+    int *omegas;
 };
 
 struct Cycloid createCycloid(int numCycles);
